:art: Migrate Aladin object from __init__ to dedicated file
